+++ cleanbcdx-goelectricbc/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.8 July 15, 2026
 – Patched ECV block styles
 – Fixed Vue bundle injection for defintions modal use
+– Updated decision date table label to "Approved since"
 
 ## 1.0.7 July 14, 2026
 – Eligible commercial vehicles block and related output